front end folder structure changed, app.jsx updated to render navbar and outlet
