+++ client/src/components/possibleComponnents/CommentForm/index.jsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useMutation } from '@apollo/client';
 
-import { ADD_COMMENT } from '../../utils/mutations';
+import { ADD_COMMENT } from '../../../utils/mutations';
 
-import Auth from '../../utils/auth';
+import Auth from '../../../utils/auth';
 
 const CommentForm = ({ thoughtId }) => {
   const [commentText, setCommentText] = useState('');
