+++ client/src/components/possibleComponnents/SingleRecipe.jsx
@@ -5,7 +5,7 @@ import { useQuery } from '@apollo/client';
 import CommentList from './CommentList';
 import CommentForm from './CommentForm';
 
-import { QUERY_SINGLE_THOUGHT } from '../utils/queries';
+import { QUERY_SINGLE_THOUGHT } from '../../utils/queries';
 
 const SingleThought = () => {
   // Use `useParams()` to retrieve value of the route parameter `:profileId`
