+++ client/src/components/possibleComponnents/ThoughtForm/index.jsx
