+++ client/src/components/possibleComponnents/ThoughtList/index.jsx
