+++ client/src/pages/SearchRecipes.jsx
@@ -1,30 +1,22 @@
 import { useState, useEffect } from 'react';
-import {
-  Container,
-  Col,
-  Form,
-  Button,
-  Card,
-  Row
-} from 'react-bootstrap';
+import { Container, Col, Form, Button, Card, Row} from 'react-bootstrap';
 import {useMutation} from '@apollo/client';
 import Auth from '../utils/auth';
 import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
 import { SAVE_BOOK } from '../utils/mutations';
+
 const SearchBooks = () => {
   const [searchedBooks, setSearchedBooks] = useState([]);
   const [searchInput, setSearchInput] = useState('');
   const [savedBookIds, setSavedBookIds] = useState(getSavedBookIds());
   const [saveBook, {error}] = useMutation(SAVE_BOOK)
 
-  // set up useEffect hook to save `savedBookIds` list to localStorage on component unmount
-  // learn more here: https://reactjs.org/docs/hooks-effect.html#effects-with-cleanup
   useEffect(() => {
    saveBookIds(savedBookIds);
    return () => {
   
   };
-}, []);
+  }, []);
 
   const handleFormSubmit = async (event) => {
     event.preventDefault();
